requirements.txt
fastapi
uvicorn[standard]
pinecone-client
sentence-transformers
python-dotenv
pinecone-client[grpc]